+++ src/main/java/university/ui/TeacherMenu.java
@@ -190,19 +190,23 @@ public class TeacherMenu {
         boolean found = false;
         Teacher teacher = null;
         while (!found) {
-            System.out.println("Введіть ідентифікатор вчителя, якого треба видалити");
+            System.out.println("Введіть ідентифікатор вчителя, якого треба видалити (нуль, щоб вийти)");
             String teacherId = scanner.nextLine();
+            if (teacherId.equals("0")) {
+                return;
+            }
 
             Optional<Teacher> optionalTeacher = teacherRepository.findById(teacherId);
             if (optionalTeacher.isPresent()) {
                 teacher = optionalTeacher.get();
                 found = true;
+                teacherService.deleteTeacher(teacher);
             } else {
                 System.out.println("Викладача з таким ID не знайдено.");
                 break;
             }
         }
-        teacherService.deleteTeacher(teacher);
+
     }
 
     protected void changeTeacher() {
